+++ HISTORY.rst
@@ -9,6 +9,7 @@ History
 * Fix: remove method in actions.leica.comment.CommentMail._recipients that was buggy and exists in the parent class (rudaporto).
 * Added a logger.warn call to track when there are no recipients in mail actions (rudaporto).
 * Fix: InternalEvent needs to implement IInternalEvent (ericof).
+* Add Slack notification to leads.PackageOrder creation (ericof).
 
 2.0.21 (2017-08-22)
 -------------------